Yöntem değişti, ProductModel'a required eklenip bir değişkene listteki değerler atandı, count saydırmada olduğu gibi burada da product'tan veriyi alıp scanbarcode ekranında işliyoruz , ek düzenlemeler yapıldı.

+++ EasyBarcodeMAU/Models/ProductItemBase.cs
@@ -1,6 +1,9 @@
 ﻿namespace EasyBarcodeMAU.Models;
     public class ProductItemBase {
-        public int Id { get; set; }
+
+    #region Properties
+    
+    public int Id { get; set; }
         public int DefterNo { get; set; }
         public DateTime DepoGirisi { get; set; }
         public string MusteriAd { get; set; }
@@ -9,6 +12,9 @@
         public int UrunAgirlik { get; set; }
         public string DepoKonum { get; set; }
         public long BarkodNo { get; set; }   
-        public int RequiredCount { get; set; }   
-        public long ReadedCount { get; set; }   
-    }
+        public long RequiredCount { get; set; }   
+        public long ReadedCount { get; set; }
+
+    #endregion
+
+}
